requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
catboost>=1.2.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Hyperparameter Optimization
optuna>=3.0.0

# Model Explainability
shap>=0.42.0

# Clustering (K-Prototypes) - Note: kmodes 0.12.2 is latest available
kmodes>=0.12.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Notebook Environment
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
ipython>=8.15.0
ipywidgets>=8.1.0

# Configuration & I/O
python-dotenv>=1.0.0
joblib>=1.3.0
PyYAML>=6.0.0

# Code Quality & Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0

# Type Checking
mypy>=1.5.0
